+++ packages/client/graph-scaffolder/package.json
@@ -4,7 +4,7 @@
   "description": "Scaffolds SVG node-and-link diagrams.",
   "license": "Apache-2.0",
   "main": "src/index.ts",
-  "types": "src/types/index.d.ts",
+  "types": "src/types/index.ts",
   "type": "module",
   "files": [
     "build/dist",
