+++ packages/client/graph-scaffolder/src/types/index.ts
@@ -49,3 +49,29 @@ export interface IGraph<V, E> {
 export type D3Selection = d3.Selection<d3.BaseType, any, null, any>;
 export type D3SelectionINode<T> = d3.Selection<d3.BaseType, INode<T>, null, any>;
 export type D3SelectionIEdge<T> = d3.Selection<d3.BaseType, IEdge<T>, null, any>;
+
+type AsyncFunction<A, O> = (args: A) => Promise<O>;
+type AsyncLayoutFunction<V, E> = AsyncFunction<IGraph<V, E>, IGraph<V, E>>;
+type LayoutFunction<V, E> = (args: IGraph<V, E>) => IGraph<V, E>;
+
+export interface Options {
+	el?: HTMLDivElement | null;
+	runLayout: AsyncLayoutFunction<any, any> | LayoutFunction<any, any>;
+
+	useZoom?: boolean;
+	zoomRange?: [number, number];
+	useStableLayout?: boolean;
+
+	// Attempt to use the same set of zoom parameters across layout changes
+	useStableZoomPan?: boolean;
+
+	// This is getting around algorithms that do not provide stand-alone routing capabilities, in
+	// which case we can internally route using A-star
+	useAStarRouting?: boolean;
+
+	// Whether to show grid
+	useGrid?: boolean;
+
+	// The css class for custom dragging event
+	dragSelector?: string;
+}
